move drizzle from desktop to drizzle package
